For AFURLConnectionOperation, modify cancel to mark the operation as cancelled regardless whether or not the operation was marked as finished.
This change helps code prevent crashes caused by the following behavior.

Object A contains a property that retains an AFURLConnectionOperation object. That AFURLConnectionOperation object may contain a callback that weakly references object A. When we release object A--causing it to get deallocated--the callback to AFURLConnectionOperation may still get called. Without giving someway for the callback to determine whether or not object A has already been released, the callback may crash. This commit gives callback the ability to determine whether or not the operation was cancelled by object A (in the event object A gets deallocated).